+++ glsl.d.ts
@@ -1,5 +1,9 @@
 // https://stackoverflow.com/questions/48741570/how-can-i-import-glsl-as-string-in-typescript
-declare module "*.glsl" {
+declare module "*.vert" {
+  const value: string;
+  export default value;
+}
+declare module "*.frag" {
   const value: string;
   export default value;
 }