+++ src/Resizer.ts
@@ -1,8 +1,8 @@
 class Resizer {
   constructor(
     container: HTMLElement,
-    camera: THREE.PerspectiveCamera = undefined,
-    renderer: THREE.WebGLRenderer = undefined
+    camera: THREE.PerspectiveCamera = undefined as any,
+    renderer: THREE.WebGLRenderer = undefined as any
   ) {
     // set initial size on load
     this.setSize(container, camera, renderer);